Fix GCC 15 format string warnings
Align format specifiers with unsigned and fixed-width types
across Bareos components so GCC 15 no longer emits
-Wformat errors.

Keep logging, SQL, and command output behavior unchanged
apart from the specifier fixes needed for a clean build.

+++ core/src/dird/date_time.cc
@@ -76,8 +76,9 @@ DateTime::DateTime(time_t time_) : time(time_)
 
 void DateTime::PrintDebugMessage(int debug_level) const
 {
-  Dmsg8(debug_level, "now = %lx: h=%d m=%d md=%d wd=%d woy=%d yday=%d\n ", time,
-        hour, month, day_of_month, day_of_week, week_of_year, day_of_year);
+  Dmsg8(debug_level, "now = %lx: h=%d m=%d md=%d wd=%d woy=%d yday=%d\n ",
+        static_cast<unsigned long>(time), hour, month, day_of_month,
+        day_of_week, week_of_year, day_of_year);
 }
 
 }  // namespace directordaemon
