+++ core/src/dird/ua_status.cc
@@ -74,7 +74,7 @@ static void StatusContentJson(UaContext* ua, StorageResource* store);
 
 inline constexpr const char OKdotstatus[] = "1000 OK .status\n";
 inline constexpr const char DotStatusJob[]
-    = "JobId=%s JobStatus=%c JobErrors=%d\n";
+    = "JobId=%s JobStatus=%c JobErrors=%" PRIu32 "\n";
 
 static void ClientStatus(UaContext* ua, ClientResource* client, char* cmd)
 {
@@ -132,7 +132,7 @@ bool DotStatusCmd(UaContext* ua, const char* cmd)
             = RecentJobResultsList::GetMostRecentJobResult();
         if (ua->AclAccessOk(Job_ACL, job.Job)) {
           ua->SendMsg(DotStatusJob, edit_int64(job.JobId, ed1), job.JobStatus,
-                      job.Errors);
+                      static_cast<uint32_t>(job.Errors));
         }
       }
     } else if (Bstrcasecmp(ua->argk[2], "header")) {
@@ -294,7 +294,7 @@ static void DoAllStatus(UaContext* ua)
     }
     if (!found) {
       unique_store[i++] = store;
-      Dmsg2(40, "Stuffing: %s:%d\n", store->address, store->SDport);
+      Dmsg2(40, "Stuffing: %s:%" PRIu32 "\n", store->address, store->SDport);
     }
   }
 
@@ -327,7 +327,7 @@ static void DoAllStatus(UaContext* ua)
     }
     if (!found) {
       unique_client[i++] = client;
-      Dmsg2(40, "Stuffing: %s:%d\n", client->address, client->FDport);
+      Dmsg2(40, "Stuffing: %s:%" PRIu32 "\n", client->address, client->FDport);
     }
   }
 
@@ -1188,11 +1188,12 @@ static void ListRunningJobs(UaContext* ua)
 
     if (ua->api) {
       BashSpaces(jcr->comment);
-      ua->SendMsg(T_("%6d\t%-6s\t%-20s\t%s\t%s\n"), jcr->JobId, level, jcr->Job,
-                  msg, jcr->comment);
+      ua->SendMsg(T_("%6" PRIu32 "\t%-6s\t%-20s\t%s\t%s\n"), jcr->JobId, level,
+                  jcr->Job, msg, jcr->comment);
       UnbashSpaces(jcr->comment);
     } else {
-      ua->SendMsg(T_("%6d %-6s  %-20s %s\n"), jcr->JobId, level, jcr->Job, msg);
+      ua->SendMsg(T_("%6" PRIu32 " %-6s  %-20s %s\n"), jcr->JobId, level,
+                  jcr->Job, msg);
       /* Display comments if any */
       if (*jcr->comment) {
         ua->SendMsg(T_("               %-30s\n"), jcr->comment);
@@ -1279,13 +1280,13 @@ static void ListTerminatedJobs(UaContext* ua)
         break;
     }
     if (ua->api) {
-      ua->SendMsg(T_("%6d\t%-6s\t%8s\t%10s\t%-7s\t%-8s\t%s\n"), je.JobId, level,
-                  edit_uint64_with_commas(je.JobFiles, b1),
+      ua->SendMsg(T_("%6" PRIu32 "\t%-6s\t%8s\t%10s\t%-7s\t%-8s\t%s\n"),
+                  je.JobId, level, edit_uint64_with_commas(je.JobFiles, b1),
                   edit_uint64_with_suffix(je.JobBytes, b2), termstat, dt,
                   JobName);
     } else {
-      ua->SendMsg(T_("%6d  %-6s %8s %10s  %-7s  %-8s %s\n"), je.JobId, level,
-                  edit_uint64_with_commas(je.JobFiles, b1),
+      ua->SendMsg(T_("%6" PRIu32 "  %-6s %8s %10s  %-7s  %-8s %s\n"),
+                  je.JobId, level, edit_uint64_with_commas(je.JobFiles, b1),
                   edit_uint64_with_suffix(je.JobBytes, b2), termstat, dt,
                   JobName);
     }
