+++ core/src/findlib/bfile.cc
@@ -1008,7 +1008,9 @@ int bopen(BareosFilePacket* bfd,
           dev_t rdev)
 {
   Dmsg4(100, "bopen: fname %s, flags %08o, mode %04o, rdev %llu\n", fname,
-        flags, (mode & ~S_IFMT), static_cast<long long unsigned>(rdev));
+        static_cast<unsigned int>(flags),
+        static_cast<unsigned int>(mode & ~S_IFMT),
+        static_cast<long long unsigned>(rdev));
 
   if (bfd->cmd_plugin && plugin_bopen) {
     Dmsg1(400, "call plugin_bopen fname=%s\n", fname);
